sidecars/live_telemetry_sidecar/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "live_telemetry_sidecar"
version = "0.1.0"
description = "Standalone telemetry backend sidecar"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websockets",
    "pyirsdk",
]

[project.scripts]
live_telemetry_sidecar = "live_telemetry_sidecar.engine:main"

[tool.setuptools.packages.find]
where = ["src"]
